refactor(peer_handler): add models.rs and struct for sender
