+++ massa-protocol-worker-2/src/connectivity.rs
@@ -52,8 +52,7 @@ pub fn start_connectivity_thread(
     let (sender_blocks_ext, receiver_blocks_ext) = unbounded();
 
     let handle = std::thread::spawn(move || {
-        let (mut peer_manager_handler, sender_peers, _sender_cmd) =
-            PeerManagementHandler::new(initial_peers);
+        let mut peer_management_handler = PeerManagementHandler::new(initial_peers);
         //TODO: Bound the channel
         // Channels network <-> handlers
         let (sender_operations, receiver_operations) = unbounded();
@@ -65,7 +64,7 @@ pub fn start_connectivity_thread(
             sender_blocks,
             sender_endorsements,
             sender_operations,
-            sender_peers,
+            sender_peers: peer_management_handler.sender.msg_sender.clone(),
             id_deserializer: U64VarIntDeserializer::new(Included(0), Included(u64::MAX)),
         };
 
@@ -109,7 +108,7 @@ pub fn start_connectivity_thread(
             select! {
                 recv(receiver) -> msg => {
                     if let Ok(ConnectivityCommand::Stop) = msg {
-                        if let Some(handle) = peer_manager_handler.thread_join.take() {
+                        if let Some(handle) = peer_management_handler.thread_join.take() {
                             handle.join().expect("Failed to join peer manager thread");
                         }
                         operation_handler.stop();
@@ -130,7 +129,7 @@ pub fn start_connectivity_thread(
                     };
                     // Get the best peers
                     {
-                        let peer_db_read = peer_manager_handler.peer_db.read();
+                        let peer_db_read = peer_management_handler.peer_db.read();
                         let best_peers = peer_db_read.index_by_newest.iter().take(nb_connection_to_try);
                         for (_timestamp, peer_id) in best_peers {
                             let peer_info = peer_db_read.peers.get(peer_id).unwrap();
