+++ massa-protocol-worker-2/src/handlers/peer_handler/tester.rs
@@ -14,7 +14,8 @@ use peernet::{
 
 use super::{
     announcement::{AnnouncementDeserializer, AnnouncementDeserializerArgs},
-    PeerInfo, SharedPeerDB,
+    models::PeerInfo,
+    SharedPeerDB,
 };
 
 #[derive(Clone)]
